jros2 conversion

+++ ihmc-high-level-behaviors/src/main/java/us/ihmc/behaviors/behaviorTree/action/actions/SceneActionDefinition.java
@@ -1,12 +1,12 @@
 package us.ihmc.behaviors.behaviorTree.action.actions;
 
-import behavior_msgs.msg.dds.SceneActionDefinitionMessage;
+import behavior_msgs.SceneActionDefinitionMessage;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.ArrayNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import gnu.trove.list.array.TIntArrayList;
 import org.yaml.snakeyaml.Yaml;
-import perception_msgs.msg.dds.YOLOv8ModelInfo;
+import perception_msgs.YOLOv8ModelInfo;
 import us.ihmc.behaviors.behaviorTree.BehaviorTreeRootNodeDefinition;
 import us.ihmc.behaviors.behaviorTree.action.ActionNodeDefinition;
 import us.ihmc.behaviors.behaviorTree.scene.BehaviorTreeSceneObjectDefinition;
@@ -307,7 +307,7 @@ public class SceneActionDefinition extends ActionNodeDefinition
       message.setSceneActionType(sceneActionType.toMessageOrdinal());
       sceneObjectDefinition.toMessage(message.getSceneObjectDefinition());
       message.setTimeout(timeout.toMessage());
-      message.setMinimumHistorySize(minimumHistorySize.toMessage());
+      message.setMinimumHistorySize((short) minimumHistorySize.toMessage());
       nominalObjectPose.toMessage(message.getNominalObjectPose());
       message.setPoseFilterAlpha(poseFilterAlpha.toMessage());
       message.setAcceptanceConfidence(acceptanceConfidence.toMessage());
